feat(product_detail): move store and slice folders to redux folder
